refactor: change order for publishing (#6)
This completely changes the order of steps we follow for publishing a
package. The goal here is to move `npm publish` as close as possible to
the beginning so that we can error-out early if it fails (e.g. 2FA
errors).

At a high level, our new steps are:

 * npm version (bump version, commit, make tag)
 * echo npm publish
   * if successful, push commit & tag
   * if failed, undo commit & tag

In the case of failure, we still provide the same granularity of error
output: we check to see if the user is logged in, if they're
owner/contributor, etc. In addition, we also provide some error output
if the user is outside of a git repo (the usual cause for `npm version`
failures).

While this refactor is done to enable supporting 2FA, this does not
itself provide OTP support. This refactor also adds TODOs for other
bugs, which will be resolved in follow-up PRs.

Issue #5