fix: require minimum npm version for otp flag
This adds an extra logic check to verify that we at least have the
minimum npm version which supports the --otp switch (as far as I can
tell, that's npm v6). Instead of attempting to `npm publish`, this
instead throws an error to tell the user to upgrade their npm
installation.

Note that this is not a new requirement; previously the `npm publish`
command would just fail and (confusingly) tell the user they need to use
the --otp switch. This change just makes the error output clearer.

Test: nvm use v7 (Node v7 comes with npm v4)
Test: npm run release:patch -- --otp=NO --release-branch=fix-require-min-npm-ver